* elf32-bfin.c (bfinfdpic_relocs_info_hash): Sprinkle casts to 	eliminate warnings.

+++ bfd/elf32-bfin.c
@@ -1284,8 +1284,8 @@ bfinfdpic_relocs_info_hash (const void *entry_)
   const struct bfinfdpic_relocs_info *entry = entry_;
 
   return (entry->symndx == -1
-	  ? entry->d.h->root.root.hash
-	  : entry->symndx + entry->d.abfd->id * 257) + entry->addend;
+	  ? (long) entry->d.h->root.root.hash
+	  : entry->symndx + (long) entry->d.abfd->id * 257) + entry->addend;
 }
 
 /* Test whether the key fields of two bfinfdpic_relocs_info entries are
